Fixed a few mismatched prototypes.
[git-p4: depot-paths = "//depot/openwatcom/": change = 24608]

+++ bld/sdk/imgedit/h/imgedit.h
@@ -240,8 +240,8 @@ void SetInitScreenColour( COLORREF colour );
 /* curclr.c */
 MRESULT CALLBACK CurrentWndProc( HWND hwnd, WPI_MSG msg, WPI_PARAM1 wparam, WPI_PARAM2 lparam );
 void CreateCurrentWnd( HWND hparent );
-void SetColour( short mousebutton, COLORREF colour, COLORREF solid, wie_clrtype type );
-COLORREF GetSelectedColour( short mousebutton, COLORREF *solid, wie_clrtype *type );
+void SetColour( int mousebutton, COLORREF colour, COLORREF solid, wie_clrtype type );
+COLORREF GetSelectedColour( int mousebutton, COLORREF *solid, wie_clrtype *type );
 void VerifyCurrentClr( COLORREF screen_colour, COLORREF inverse_colour );
 void SetCurrentNumColours( int colour_count );
 void ChangeCurrentColour( void );
